NewModelFactory => ModelFactory, deleted old stuff
