+++ ModelFactories/ModelFactories.csproj
@@ -6,7 +6,7 @@
 	<Nullable>enable</Nullable>
 
 	<PackageId>ModelFactories.Local</PackageId>
-	<Version>0.0.1</Version>
+        <Version>0.1.0</Version>
 	<Authors>Larsmbergvall</Authors>
         <AssemblyVersion>0.1.0</AssemblyVersion>
     </PropertyGroup>
